refactor: refactor class name
